Replace psa-attestation/lib/mbed-crypto with mbedtls from rust-psa-crypto.
The crate rust-psa-crypto is slightly modified from upstream, primarily
to configure mbedtls differently: "baremetal" with entropy obtained with
the C-library function getrandom.

Also, rust-psa-crypto/Cargo.toml is removed to avoid nested workspaces.

This uses a different version of the C library with a slightly different
API. In particular, key handles are now uint32_t instead of uint16_t.
This accounts for most of the changes in psa-attestation and t_cose.

There are various IceCap changes to enable building rust-psa-crypto.
These include adding a bogus implementation of getrandom()!

+++ psa-attestation/include/psa/initial_attestation.h
@@ -51,9 +51,9 @@ typedef int32_t psa_status_t;
 
 psa_status_t psa_initial_attest_load_key(uint8_t const *private_key,
                                          size_t private_key_size,
-                                         uint16_t *p_key_handle);
+                                         uint32_t *p_key_handle);
 
-psa_status_t psa_initial_attest_remove_key(uint16_t key_handle);
+psa_status_t psa_initial_attest_remove_key(uint32_t key_handle);
 
 /**
  * The list of fixed claims in the initial attestation token is still evolving,
